prediction/python_scripts/requirements.txt
pandas
pyyaml
scikit-learn
joblib
